Updated LeetCode Files

+++ LeetCode/Hard/Max Chunks To Make Sorted II.cpp
@@ -4,7 +4,7 @@ using namespace std;
 class Solution {
 public:
     int maxChunksToSorted(vector<int>& arr) {
-        int sum1 = 0, sum2 = 0, ans = 0;
+        long sum1 = 0, sum2 = 0, ans = 0;
         vector<int> t = arr;
         sort(begin(t), end(t));
         for (int i = 0; i < arr.size(); ++i){
